Consolidate assignment vertical slice
refs: MBL-14386
affects: none
release note: none
